word2blank/word2blank/requirements.txt
gensim
pudb
click
tabulate
terminalplot
progressbar2
